Fix pawn promotion notation interpretation, fix takeback overwriting, add tool visibility control, fix engine color initialization after start.

+++ constants.cpp
@@ -42,4 +42,4 @@ const Board starting_board = {
      Piece(Color::White, 'K'), Piece(Color::White, 'B'),
      Piece(Color::White, 'N'), Piece(Color::White, 'R')}};
 
-const size_t adaptive_off_threshold = 20;
+const double adaptive_off_threshold = 20.0;
