+++ constants.h
@@ -13,4 +13,4 @@ extern const int rook_moves[4];
 extern const Board starting_board;
 extern const int mate;
 
-extern const size_t adaptive_off_threshold;
+extern const double adaptive_off_threshold;
